[H2-272] change to allow for both duration and iterations
If duration <= 0, we will run the number of iterations specified.

If duration > 0, we will run at least the duration, but if iterations
take longer, we will still complete them

If iterations <= 0, we will run for just for the amount of duration
specified

If iterations complete before duration expires, that is it. Nothing else
to do.

So if you just want to run `x` amount of time, only specify duration
option.